chore：move Property.py to apex/core/property/
