+++ apex/core/property/Interstitial.py
@@ -3,7 +3,6 @@ import json
 import logging
 import os
 import re
-import copy
 
 import numpy as np
 from monty.serialization import dumpfn, loadfn
@@ -12,7 +11,7 @@ from pymatgen.core.structure import Structure
 
 import apex.core.calculator.lib.abacus as abacus
 import apex.core.calculator.lib.lammps as lammps
-from apex.core.Property import Property
+from apex.core.property.Property import Property
 from apex.core.refine import make_refine
 from apex.core.reproduce import make_repro, post_repro
 from apex.core.structure import StructureInfo
