+++ apex/core/property/Surface.py
@@ -12,7 +12,7 @@ from pymatgen.core.surface import generate_all_slabs
 
 import apex.core.calculator.lib.abacus as abacus
 import apex.core.calculator.lib.vasp as vasp
-from apex.core.Property import Property
+from apex.core.property.Property import Property
 from apex.core.refine import make_refine
 from apex.core.reproduce import make_repro, post_repro
 from dflow.python import upload_packages
