+++ apex/core/property/Vacancy.py
@@ -10,7 +10,7 @@ from pymatgen.analysis.defects.generators import VacancyGenerator
 from pymatgen.core.structure import Structure
 
 import apex.core.calculator.lib.abacus as abacus
-from apex.core.Property import Property
+from apex.core.property.Property import Property
 from apex.core.refine import make_refine
 from apex.core.reproduce import make_repro, post_repro
 from dflow.python import upload_packages
